Split Input recording and update-helper god classes under 500 lines
Extract Class on InputReplayer, InputRecorder, and InputSystemUpdateHelper so replay/record orchestration stays thin while event processing, frame capture, and runtime update waiting each have one responsibility.
