+++ Packages/src/Editor/FirstPartyTools/Common/InputRecording/InputRecorderFrameCapture.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: db2f63be712542fb816e560aa10cbf7a
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
