+++ Packages/src/Editor/FirstPartyTools/Common/InputRecording/InputRecordingVectorFormat.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4816c504fbbd4a0897ed40f554399b3f
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
