+++ Packages/src/Editor/FirstPartyTools/Common/InputRecording/InputReplayUiController.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 9e37633930b7478b97555a4524e2d452
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
