+++ Packages/src/Editor/FirstPartyTools/Common/InputSystem/InputSimulationWaitTypes.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c436823b41f74b9aa0d56e2d07493671
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
