+++ Packages/src/Editor/FirstPartyTools/Common/InputSystem/InputSystemMainThreadAwaitable.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 1779f2469cd44bc2b63c7d68c9d6cbe7
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
