+++ Packages/src/Editor/FirstPartyTools/Common/InputSystem/InputSystemRuntimeFrameWaiter.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: e495c703ea7e4da199307925db350077
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
