+++ Packages/src/Editor/FirstPartyTools/Common/InputSystem/InputSystemUpdateSubscription.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 4239026933f94ca286c891d4f8b9fe2b
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
